Fix weird type issue
